requirements.txt
argparse
sentence-transformers
langchain
langchain-community
huggingface_hub
transformers
PyPDF2
accelerate 
bitsandbytes 
faiss-gpu 
openpyxl 
pacmap 
datasets 
ragatouille 
gdown
instructor
google-generativeai
pydantic
deepeval
zipfile